Update README.md with API endpoint documentation
This update replaced the old README.md content with detailed explanations about using API endpoints. The updated instructions cover everything from accessing the Swagger-UI to the details of POST requests for eva-data and eva-cv. Users are now provided with an understanding of each field required in the request body for both endpoints.